Moved descriptor set data to Shader - Descriptor set data moved from pipeline to shader class - Changed Timer elapsed time format from float to double

+++ WingnutLib/src/Core/Timer.h
@@ -19,11 +19,11 @@ namespace Wingnut
 			m_StartTime = std::chrono::steady_clock::now();
 		}
 
-		float ElapsedTime()
+		double ElapsedTime()
 		{
 			std::chrono::steady_clock::time_point t2 = std::chrono::steady_clock::now();
 
-			return (float)std::chrono::duration_cast<std::chrono::milliseconds>(t2 - m_StartTime).count() / 1000.0 * 1000.0;
+			return std::chrono::duration_cast<std::chrono::milliseconds>(t2 - m_StartTime).count() / 1000.0 * 1000.0;
 		}
 
 	private:
